requirements.txt
torch==2.8.0
transformers==4.56.2
peft==0.17.1
datasets==4.1.1
trl==0.23.0
accelerate==1.10.1
numpy==2.3.3
huggingface-hub==0.35.1
tokenizers==0.22.1
safetensors==0.6.2
tqdm==4.67.1
requests==2.32.5
packaging==25.0
pyyaml==6.0.3
regex==2025.9.18
filelock==3.19.1
pillow==10.4.0
timm>=1.0.22